test(UserManager): Overhaul UserManager tests
Completely refactor large sections of generated UserManager tests, the
User model, and the UserManager model.

This is a required change; our existing validation, encryption, and
hashing of attributes in requests means that Phoenix assumptions
regarding a default `context` are incorrect. This commit alters the
tests to preserve their purpose, but also _work_.